requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
openai
psycopg2-binary
pdfplumber
unstructured
python-dotenv
tqdm
pillow
pillow_heif
matplotlib
PyMuPDF
asyncpg
langchain
langchain_community
PyPDF2
langchain-openai
websockets
scikit-learn